+++ src-docs/src/views/tables/selection/selection.js
@@ -203,6 +203,9 @@ export class Table extends Component {
       onSelectionChange: this.onSelectionChange,
     };
 
+    const getAriaLabel = (item = {}) =>
+      `${item.firstName || ''} ${item.lastName || ''}`;
+
     return (
       <Fragment>
         {deleteButton}
@@ -214,6 +217,7 @@ export class Table extends Component {
           sorting={sorting}
           isSelectable={true}
           selection={selection}
+          getAriaLabel={getAriaLabel}
           onChange={this.onTableChange}
         />
       </Fragment>